+++ src/main/java/com/google/firebase/auth/internal/FirebaseTokenFactory.java
@@ -16,64 +16,56 @@
 
 package com.google.firebase.auth.internal;
 
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Preconditions.checkNotNull;
+
 import com.google.api.client.json.GenericJson;
 import com.google.api.client.json.JsonFactory;
-import com.google.api.client.json.gson.GsonFactory;
 import com.google.api.client.json.webtoken.JsonWebSignature;
+import com.google.api.client.util.Base64;
 import com.google.api.client.util.Clock;
-import com.google.common.base.Preconditions;
+import com.google.api.client.util.StringUtils;
 
+import com.google.common.base.Strings;
+import com.google.firebase.FirebaseApp;
 import java.io.IOException;
-import java.security.GeneralSecurityException;
-import java.security.PrivateKey;
 import java.util.Collection;
 import java.util.Map;
 
 /**
- * Provides helper methods to simplify the creation of FirebaseCustomAuthTokens.
+ * Provides helper methods to simplify the creation of Firebase custom auth tokens.
  *
  * <p>This class is designed to hide underlying implementation details from a Firebase developer.
  */
 public class FirebaseTokenFactory {
 
-  private static FirebaseTokenFactory instance;
+  private final JsonFactory jsonFactory;
+  private final Clock clock;
+  private final CryptoSigner signer;
 
-  private JsonFactory factory;
-  private Clock clock;
-
-  public FirebaseTokenFactory(JsonFactory factory, Clock clock) {
-    this.factory = factory;
-    this.clock = clock;
+  FirebaseTokenFactory(JsonFactory jsonFactory, Clock clock, CryptoSigner signer) {
+    this.jsonFactory = checkNotNull(jsonFactory);
+    this.clock = checkNotNull(clock);
+    this.signer = checkNotNull(signer);
   }
 
-  public static FirebaseTokenFactory getInstance() {
-    if (null == instance) {
-      instance = new FirebaseTokenFactory(new GsonFactory(), Clock.SYSTEM);
-    }
-
-    return instance;
-  }
-
-  public String createSignedCustomAuthTokenForUser(String uid, String issuer, PrivateKey privateKey)
-      throws GeneralSecurityException, IOException {
-    return createSignedCustomAuthTokenForUser(uid, null, issuer, privateKey);
+  String createSignedCustomAuthTokenForUser(String uid) throws IOException {
+    return createSignedCustomAuthTokenForUser(uid, null);
   }
 
   public String createSignedCustomAuthTokenForUser(
-      String uid, Map<String, Object> developerClaims, String issuer, PrivateKey privateKey)
-      throws GeneralSecurityException, IOException {
-    Preconditions.checkState(uid != null, "Uid must be provided.");
-    Preconditions.checkState(issuer != null && !"".equals(issuer), "Must provide an issuer.");
-    Preconditions.checkState(uid.length() <= 128, "Uid must be shorter than 128 characters.");
+      String uid, Map<String, Object> developerClaims) throws IOException {
+    checkArgument(!Strings.isNullOrEmpty(uid), "Uid must be provided.");
+    checkArgument(uid.length() <= 128, "Uid must be shorter than 128 characters.");
 
     JsonWebSignature.Header header = new JsonWebSignature.Header().setAlgorithm("RS256");
 
-    long issuedAt = clock.currentTimeMillis() / 1000;
+    final long issuedAt = clock.currentTimeMillis() / 1000;
     FirebaseCustomAuthToken.Payload payload =
         new FirebaseCustomAuthToken.Payload()
             .setUid(uid)
-            .setIssuer(issuer)
-            .setSubject(issuer)
+            .setIssuer(signer.getAccount())
+            .setSubject(signer.getAccount())
             .setAudience(FirebaseCustomAuthToken.FIREBASE_AUDIENCE)
             .setIssuedAtTimeSeconds(issuedAt)
             .setExpirationTimeSeconds(issuedAt + FirebaseCustomAuthToken.TOKEN_DURATION_SECONDS);
@@ -90,7 +82,24 @@ public class FirebaseTokenFactory {
       jsonObject.putAll(developerClaims);
       payload.setDeveloperClaims(jsonObject);
     }
+    return signPayload(header, payload);
+  }
 
-    return JsonWebSignature.signUsingRsaSha256(privateKey, factory, header, payload);
+  private String signPayload(JsonWebSignature.Header header,
+      FirebaseCustomAuthToken.Payload payload) throws IOException {
+    String headerString = Base64.encodeBase64URLSafeString(jsonFactory.toByteArray(header));
+    String payloadString = Base64.encodeBase64URLSafeString(jsonFactory.toByteArray(payload));
+    String content = headerString + "." + payloadString;
+    byte[] contentBytes = StringUtils.getBytesUtf8(content);
+    String signature = Base64.encodeBase64URLSafeString(signer.sign(contentBytes));
+    return content + "." + signature;
+  }
+
+  public static FirebaseTokenFactory fromApp(
+      FirebaseApp firebaseApp, Clock clock) throws IOException {
+    return new FirebaseTokenFactory(
+        firebaseApp.getOptions().getJsonFactory(),
+        clock,
+        CryptoSigners.getCryptoSigner(firebaseApp));
   }
 }