+++ src/test/java/com/google/firebase/auth/FirebaseAuthIT.java
@@ -35,11 +35,16 @@ import com.google.api.client.json.JsonObjectParser;
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutureCallback;
 import com.google.api.core.ApiFutures;
+import com.google.auth.ServiceAccountSigner;
+import com.google.auth.oauth2.AccessToken;
+import com.google.auth.oauth2.GoogleCredentials;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.BaseEncoding;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import com.google.firebase.ImplFirebaseTrampolines;
 import com.google.firebase.auth.UserRecord.CreateRequest;
 import com.google.firebase.auth.UserRecord.UpdateRequest;
 import com.google.firebase.auth.hash.Scrypt;
@@ -365,6 +370,31 @@ public class FirebaseAuthIT {
   }
 
   @Test
+  public void testCustomTokenWithIAM() throws Exception {
+    FirebaseApp masterApp = IntegrationTestUtils.ensureDefaultApp();
+    GoogleCredentials credentials = ImplFirebaseTrampolines.getCredentials(masterApp);
+    AccessToken token = credentials.getAccessToken();
+    if (token == null) {
+      token = credentials.refreshAccessToken();
+    }
+    FirebaseOptions options = new FirebaseOptions.Builder()
+        .setCredentials(GoogleCredentials.of(token))
+        .setServiceAccountId(((ServiceAccountSigner) credentials).getAccount())
+        .setProjectId(IntegrationTestUtils.getProjectId())
+        .build();
+    FirebaseApp customApp = FirebaseApp.initializeApp(options, "tempApp");
+    try {
+      FirebaseAuth auth = FirebaseAuth.getInstance(customApp);
+      String customToken = auth.createCustomTokenAsync("user1").get();
+      String idToken = signInWithCustomToken(customToken);
+      FirebaseToken decoded = auth.verifyIdTokenAsync(idToken).get();
+      assertEquals("user1", decoded.getUid());
+    } finally {
+      customApp.delete();
+    }
+  }
+
+  @Test
   public void testVerifyIdToken() throws Exception {
     String customToken = auth.createCustomTokenAsync("user2").get();
     String idToken = signInWithCustomToken(customToken);
@@ -541,5 +571,4 @@ public class FirebaseAuthIT {
       assertEquals("uid-already-exists", ((FirebaseAuthException) e.getCause()).getErrorCode());
     }
   }
-
 }