+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- [added] Implemented the ability to create custom tokens without
+  service account credentials.
+- [added] Added the `setServiceAccount()` method to the
+  `FirebaseOptions.Builder` API.
 - [added] The SDK can now read the Firebase/GCP project ID from both
   `GCLOUD_PROJECT` and `GOOGLE_CLOUD_PROJECT` environment variables.